add support for running example* functions and reorganize vscode extension
